backend/requirements.txt
# backend/requirements.txt

fastapi
h5py<3.0.0
numpy
opencv-python
pillow
python-multipart
uvicorn
tensorflow==2.0.0
ISR==2.2.0
torch
torchvision
scipy
matplotlib
scikit-image
imgaug
keras==2.3.1
cython
IPython